new recovery phrase layout
